+++ forms/RoomDetail.cs
@@ -199,8 +199,13 @@ namespace Project.Forms {
             Program app = Program.GetInstance();
             RoomService roomService = app.GetService<RoomService>("rooms");
 
+            // Ask for confirmation
+            if (!GuiHelper.ShowConfirm("Weet je zeker dat je deze zaal wilt verwijderen?")) {
+                return;
+            }
+
             // Delete room
-            if(!roomService.DeleteRoom(room)) {
+            if (!roomService.DeleteRoom(room)) {
                 GuiHelper.ShowError("Kon zaal niet verwijderen");
                 return;
             }
